requirements.txt
lxml==4.5.0
requests==2.23.0
pytube3==9.6.4
ffmpeg-python==0.2.0